updat linux sdk

+++ ali_face_engine_sdk/Darwin/include/FaceEngine.h
@@ -41,11 +41,14 @@ namespace ali_face_engine {
 
     DLL_API void setPersistencePath(const char *path);
 
+    //version>=3.3.0
     DLL_API void setConfigString(const char *key, const char *value);
 
+    //version>=3.3.0
     DLL_API void setConfigInt(const char *key, const int value);
 
-    DLL_API void setConfigFloat(const char *key, const int value);
+    //version>=3.3.0
+    DLL_API void setConfigFloat(const char *key, const float value);
 
     //face
     DLL_API Face *newFace();
